pyproject.toml
[project]
name = "braintip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.26",
    "langchain-groq>=0.3.2",
    "langchain-ollama>=0.3.3",
    "ollama>=0.5.1",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.3",
]
